refactor: clean up authUser class
