--- src/resources/auth/middleware/OrganizationMember.ts
@@ -1,6 +1,0 @@
-export interface OrganizationMember {
-	identifier: string;
-	organizationIdentifier?: string;
-	role?: string;
-	permissionFlags: number;
-}
