+++ src/resources/auth/strategies/AuthBearerJWTStrategy.ts
@@ -14,11 +14,11 @@ export class AuthBearerJWTStrategy {
 		if (!this.strategy) {
 			this.strategy = new bearerStrategy.Strategy(async function verify(token, done) {
 				try {
-					const user = jwt.verify(token, jwtSecret);
-					if (!user) {
+					const authUser = jwt.verify(token, jwtSecret);
+					if (!authUser) {
 						return done(null, false);
 					}
-					return done(null, user);
+					return done(null, authUser);
 				} catch (err) {
 					return done(null, false);
 				}
