+++ src/schemas/kulturdaten.berlin.openapi.generated.yml
@@ -2273,6 +2273,28 @@ components:
         website:
           type: string
           example: https://example.com/
+        curator:
+          type: object
+          properties:
+            referenceType:
+              type: string
+            referenceId:
+              type: string
+            referenceLabel:
+              type: object
+              additionalProperties:
+                type: string
+              description: >-
+                A string that can be translated into multiple languages, e.g. {
+                "de": "Konzert", "en": "concert" }
+              example:
+                de: Konzert
+                en: concert
+          required:
+            - referenceType
+            - referenceId
+          additionalProperties: false
+          description: The curating organization of this attraction.
         inLanguages:
           type: array
           description: List of languages this object is available in.
@@ -3868,6 +3890,29 @@ components:
                       website:
                         type: string
                         example: https://example.com/
+                      curator:
+                        type: object
+                        properties:
+                          referenceType:
+                            type: string
+                          referenceId:
+                            type: string
+                          referenceLabel:
+                            type: object
+                            additionalProperties:
+                              type: string
+                            description: >-
+                              A string that can be translated into multiple
+                              languages, e.g. { "de": "Konzert", "en": "concert"
+                              }
+                            example:
+                              de: Konzert
+                              en: concert
+                        required:
+                          - referenceType
+                          - referenceId
+                        additionalProperties: false
+                        description: The curating organization of this attraction.
                       inLanguages:
                         type: array
                         description: List of languages this object is available in.
@@ -4161,6 +4206,29 @@ components:
                       website:
                         type: string
                         example: https://example.com/
+                      curator:
+                        type: object
+                        properties:
+                          referenceType:
+                            type: string
+                          referenceId:
+                            type: string
+                          referenceLabel:
+                            type: object
+                            additionalProperties:
+                              type: string
+                            description: >-
+                              A string that can be translated into multiple
+                              languages, e.g. { "de": "Konzert", "en": "concert"
+                              }
+                            example:
+                              de: Konzert
+                              en: concert
+                        required:
+                          - referenceType
+                          - referenceId
+                        additionalProperties: false
+                        description: The curating organization of this attraction.
                       inLanguages:
                         type: array
                         description: List of languages this object is available in.
@@ -4313,6 +4381,28 @@ components:
                 website:
                   type: string
                   example: https://example.com/
+                curator:
+                  type: object
+                  properties:
+                    referenceType:
+                      type: string
+                    referenceId:
+                      type: string
+                    referenceLabel:
+                      type: object
+                      additionalProperties:
+                        type: string
+                      description: >-
+                        A string that can be translated into multiple languages,
+                        e.g. { "de": "Konzert", "en": "concert" }
+                      example:
+                        de: Konzert
+                        en: concert
+                  required:
+                    - referenceType
+                    - referenceId
+                  additionalProperties: false
+                  description: The curating organization of this attraction.
                 inLanguages:
                   type: array
                   description: List of languages this object is available in.
@@ -7283,6 +7373,9 @@ components:
                         type: number
                       exp:
                         type: number
+                    required:
+                      - identifier
+                    additionalProperties: true
                 required:
                   - token
             user:
