+++ src/schemas/models/AccessToken.yml
@@ -3,20 +3,6 @@ properties:
   token:
     type: string
   decodedToken:
-    type: object
-    properties:
-      identifier: 
-        type: string
-      organizationIdentifier: 
-        type: string
-        nullable: true
-      role: 
-        $ref: "Role.yml"
-      permissionFlags: 
-        type: number
-      iat:
-        type: number
-      exp: 
-        type: number
+    $ref: "AuthUser.yml"
 required:
   - token
